refactor(faq): move components into features folder
